backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17
pillow>=11.0.0
python-pptx>=1.0.2
httpx>=0.28.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
